Create bb.cu
(The following is to be added as info for b&b description. More detail can be added if needed and the other section of parentheses can be included but that's up to you) 

This Branch and Bound program simulates the Knapsack problem in CUDA to search for an optimal solution given a set of 30 items with their respective weights and values. (Node pruning in this program is currently not functioning as intended, resulting in a slower program which is to be fixed soon). To build the program the following commands are to be run: 

nvcc bb.cu -o bb -I/usr/inlcude/c++/10
./bb